+++ toxcore/Messenger.h
@@ -616,7 +616,7 @@ int m_set_usertyping(Messenger *m, int32_t friendnumber, uint8_t is_typing);
  * returns 0 if friend is not typing.
  * returns 1 if friend is typing.
  */
-uint8_t m_get_istyping(const Messenger *m, int32_t friendnumber);
+int m_get_istyping(const Messenger *m, int32_t friendnumber);
 
 /* Sets whether we send read receipts for friendnumber.
  * This function is not lazy, and it will fail if yesno is not (0 or 1).
